\name{compute_gradient_coordinate}
\alias{compute_gradient_coordinate}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Finite difference function
}
\description{
	Comute the derivative of a single coordinate based on finite difference approximation.
}

\keyword{ Optimization functions }

