\name{pauwels2014-package}
\alias{pauwels2014-package}
\alias{pauwels2014}
\docType{package}
\title{
Reproduce numerical experiments
}
\description{
This package gathers functions and pre-computed results to reproduce the results presented in a research paper (see vignette). Scripts are given to compute all results from scratch or to draw pictures based on pre-computed results. Adapting this to other networks would require quite an amount of work. The vignette gives practical examples and points out the functions to modify in order to adapt to other settings. 
}
\details{
\tabular{ll}{
Package: \tab pauwels-2014\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-08-20\cr
License: \tab GPL-3\cr
}
Makes extensive use of ode solver from package \code{deSolve}. The vignette provides detailed examples.
}
\author{
Edouard Pauwels
}
\references{
	Edouard Pauwels, Christian Lajaunie, and Jean-Philippe Vert. A bayesian active learning strategy
	for sequential experimental design in systems biology. BMC Systems Biology, 2014. To appear.
  
	Karline Soetaert, Thomas Petzoldt, R. Woodrow Setzer (2010).  Solving
  Differential Equations in R: Package deSolve Journal of Statistical
  Software, 33(9), 1--25.  \url{http://www.jstatsoft.org/v33/i09/	
	}
}
\keyword{ package }

