\name{reverse_params}
\alias{reverse_params}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{
	Transform log space parameters back to the original space
}
\description{
	Least square minimization to map back log transformed parameters to their original values
}

\keyword{ Optimization functions }

